+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 
+## Unreleased
+
+- Add `Alphabet.crockfordBase32` and `Alphabet.base58`
+
 ## 2.0.1
 
 - Update pubspec topics description and repo link